Split the SQLDatabase protocol into its own file. Group SQLiteConnnection's async methods together.
